+++ src/Oro/Bundle/DataGridBundle/Resources/public/js/datagrid/grid.js
@@ -106,6 +106,10 @@ define(function(require) {
         /** @property {orodatagrid.datagrid.column.ActionColumn} */
         actionsColumn: ActionColumn,
 
+        selectRowCell: SelectRowCell,
+
+        selectAllHeaderCell: SelectAllHeaderCell,
+
         /** @property true when no one column configured to be shown in th grid */
         noColumnsFlag: false,
 
@@ -644,8 +648,8 @@ define(function(require) {
                 sortable: false,
                 editable: false,
                 manageable: false,
-                cell: SelectRowCell,
-                headerCell: SelectAllHeaderCell,
+                cell: this.selectRowCell,
+                headerCell: this.selectAllHeaderCell,
                 order: -Infinity
             });
             return column;
@@ -748,13 +752,13 @@ define(function(require) {
                     order: 700
                 });
                 this.on('appearanceChanged', function(key, options) {
-                    const item = _.findWhere(action.launcherInstanse.items,
+                    const item = _.findWhere(action.launcherInstance.items,
                         {key: options.type, id: options.id || 'by_type'});
                     if (!item) {
                         throw new Error('Could not find corresponding launcher item');
                     }
-                    action.launcherInstanse.selectedItem = item;
-                    action.launcherInstanse.render();
+                    action.launcherInstance.selectedItem = item;
+                    action.launcherInstance.render();
                 });
                 actions.push(action);
             }
@@ -901,14 +905,13 @@ define(function(require) {
          * @private
          */
         _listenToCollectionEvents: function() {
-            this.listenTo(this.collection, 'request', function(model, xhr) {
-                this._beforeRequest();
-                const self = this;
+            this.listenTo(this.collection, 'request', function(model, xhr, options = {}) {
+                this._beforeRequest(options);
                 const always = xhr.always;
-                xhr.always = function(...args) {
-                    always.apply(this, args);
-                    if (!self.disposed) {
-                        self._afterRequest(this);
+                xhr.always = (...args) => {
+                    always.apply(xhr, args);
+                    if (!this.disposed) {
+                        this._afterRequest(xhr, options);
                     }
                 };
             });
@@ -1059,10 +1062,12 @@ define(function(require) {
             this.renderLoadingMask();
 
             this.delegateEvents();
-            this.listenTo(this.collection, 'reset', this.renderNoDataBlock);
+            this.listenTo(this.collection, 'reset remove sync', this.renderNoDataBlock);
 
             this._deferredRender();
-            this.initLayout().always(_.bind(function() {
+            this.initLayout({
+                datagrid: this
+            }).always(_.bind(function() {
                 this.rendered = true;
                 /**
                  * Backbone event. Fired when the grid has been successfully rendered.
@@ -1235,9 +1240,14 @@ define(function(require) {
          *
          * @private
          */
-        _beforeRequest: function() {
+        _beforeRequest: function(options = {}) {
+            const {toggleLoading = true} = options;
             this.requestsCount += 1;
-            this.showLoading();
+
+            if (toggleLoading) {
+                this.showLoading();
+            }
+            this.lockToolBar();
         },
 
         /**
@@ -1245,15 +1255,20 @@ define(function(require) {
          *
          * @private
          */
-        _afterRequest: function(jqXHR) {
+        _afterRequest: function(jqXHR, options = {}) {
             const json = jqXHR.responseJSON || {};
+            const {toggleLoading = true} = options;
+
             if (json.metadata) {
                 this._processLoadedMetadata(json.metadata);
             }
 
             this.requestsCount -= 1;
             if (this.requestsCount === 0) {
-                this.hideLoading();
+                if (toggleLoading) {
+                    this.hideLoading();
+                }
+                this.unlockToolBar();
                 /**
                  * Backbone event. Fired when data for grid has been successfully rendered.
                  * @event grid_load:complete
@@ -1276,19 +1291,31 @@ define(function(require) {
         },
 
         /**
-         * Show loading mask and disable toolbar
+         * Show loading mask
          */
         showLoading: function() {
             this.loadingMask.show();
+        },
+
+        /**
+         * Disable toolbar
+         */
+        lockToolBar: function() {
             this.callToolbar('disable');
             this.trigger('disable');
         },
 
         /**
-         * Hide loading mask and enable toolbar
+         * Hide loading mask
          */
         hideLoading: function() {
             this.loadingMask.hide();
+        },
+
+        /**
+         * Enable toolbar
+         */
+        unlockToolBar: function() {
             this.callToolbar('enable');
             this.trigger('enable');
         },
@@ -1308,11 +1335,27 @@ define(function(require) {
          *
          * @private
          */
-        _onRemove: function(model, reset) {
+        _onRemove: function(model, collection, options = {}) {
             mediator.trigger('datagrid:beforeRemoveRow:' + this.name, model);
 
-            if (reset !== false) {
-                this.collection.fetch({reset: true});
+            if (collection) {
+                const fetchKeys = ['mode', 'parse', 'reset', 'wait', 'uniqueOnly',
+                    'add', 'remove', 'merge', 'toggleLoading'];
+                let fetchOptions = {
+                    reset: true,
+                    alreadySynced: true // prevents recursion update
+                };
+                let params = _.pick(collection.options, fetchKeys);
+
+                if (collection.options.parseResponseOptions) {
+                    params = _.extend( params, _.pick(collection.options.parseResponseOptions(), fetchKeys));
+                }
+
+                fetchOptions = _.extend(fetchOptions, params, _.pick(options, fetchKeys));
+
+                if (!options.alreadySynced) {
+                    this.collection.fetch(fetchOptions);
+                }
             }
 
             mediator.trigger('datagrid:afterRemoveRow:' + this.name);