+++ src/Oro/Bundle/DataGridBundle/Tests/Unit/Extension/InlineEditing/InlineEditingExtensionTest.php
@@ -143,6 +143,7 @@ class InlineEditingExtensionTest extends \PHPUnit\Framework\TestCase
                 'enable' => true,
                 'entity_name' => $entityName,
                 'behaviour' => 'enable_all',
+                'mobile_enabled' => false,
                 'save_api_accessor' => [
                     'route' => 'oro_api_patch_entity_data',
                     'http_method' => 'PATCH',
@@ -200,6 +201,8 @@ class InlineEditingExtensionTest extends \PHPUnit\Framework\TestCase
     }
 
     /**
+     * @return \Closure
+     *
      * @SuppressWarnings(PHPMD.CyclomaticComplexity)
      */
     protected function getProcessConfigsCallBack()