+++ CHANGELOG.md
@@ -2,6 +2,19 @@ Please refer first to [UPGRADE.md](UPGRADE.md) for the most important items that
 
 The current file describes significant changes in the code that may affect the upgrade of your customizations.
 
+## 4.2.0-rc
+
+### Added
+
+#### LayoutBundle
+* Added `is_xml_http_request` option to the Layout context which lets you know if the current request is an ajax request.
+* Added two new options `onLoadingCssClass` and `disableControls` to the `layout_subtree_update` block configuration.
+
+### Removed
+
+### SyncBundle
+* Removed long-unused the `orosync/js/content/grid-builder` component from the layout updates.
+
 ## 4.2.0-alpha.3
 
 ### Changed